feat(nexus): added store desrializer trait
